+++ apps/web/next.config.js
@@ -13,7 +13,7 @@ const nextConfig = {
   i18n: i18n,
   productionBrowserSourceMaps: true,
   output: "standalone",
-  transpilePackages: ["@rallly/emails", "@rallly/database"],
+  transpilePackages: ["@rallly/emails", "@rallly/database", "@rallly/backend"],
   webpack(config) {
     config.module.rules.push({
       test: /\.svg$/,
