+++ packages/backend/next/index.ts
@@ -1,4 +1,3 @@
-import { IronSessionOptions, sealData, unsealData } from "iron-session";
 import { withIronSessionApiRoute, withIronSessionSsr } from "iron-session/next";
 import {
   GetServerSideProps,
@@ -6,54 +5,12 @@ import {
   NextApiHandler,
 } from "next";
 
-import { createSSGHelperFromContext } from "../server/context";
+import { sessionOptions } from "@/session";
+import { createSSGHelperFromContext } from "@/trpc/context";
 
-export const sessionOptions: IronSessionOptions = {
-  password: process.env.SECRET_PASSWORD ?? "",
-  cookieName: "rallly-session",
-  cookieOptions: {
-    secure: process.env.NODE_ENV === "production",
-  },
-  ttl: 60 * 24 * 30, // 30 days
-};
-
-export type RegistrationTokenPayload = {
-  name: string;
-  email: string;
-  code: string;
-};
-
-export type LoginTokenPayload = {
-  userId: string;
-  code: string;
-};
-
-export type RegisteredUserSession = {
-  isGuest: false;
-  id: string;
-  name: string;
-  email: string;
-};
-
-export type GuestUserSession = {
-  isGuest: true;
-  id: string;
-};
-
-export type DisableNotificationsPayload = {
-  pollId: string;
-  watcherId: number;
-};
-
-export type UserSession = GuestUserSession | RegisteredUserSession;
-
-export function withSessionRoute(handler: NextApiHandler) {
-  return withIronSessionApiRoute(handler, sessionOptions);
-}
-
-export const composeGetServerSideProps = (
+export function composeGetServerSideProps(
   ...fns: GetServerSideProps[]
-): GetServerSideProps => {
+): GetServerSideProps {
   return async (ctx) => {
     const res = { props: {} };
     for (const getServerSideProps of fns) {
@@ -71,7 +28,11 @@ export const composeGetServerSideProps = (
 
     return res;
   };
-};
+}
+
+export function withSessionRoute(handler: NextApiHandler) {
+  return withIronSessionApiRoute(handler, sessionOptions);
+}
 
 /**
  * Require user to be logged in if AUTH_REQUIRED is true
@@ -130,28 +91,3 @@ export function withSessionSsr(
     return res;
   }, sessionOptions);
 }
-
-export const decryptToken = async <P extends Record<string, unknown>>(
-  token: string,
-): Promise<P | null> => {
-  const payload = await unsealData(token, {
-    password: sessionOptions.password,
-  });
-  if (Object.keys(payload).length === 0) {
-    return null;
-  }
-
-  return payload as P;
-};
-
-export const createToken = async <T extends Record<string, unknown>>(
-  payload: T,
-  options?: {
-    ttl?: number;
-  },
-) => {
-  return await sealData(payload, {
-    password: sessionOptions.password,
-    ttl: options?.ttl ?? 60 * 15, // 15 minutes
-  });
-};
