+++ packages/backend/package.json
@@ -5,12 +5,13 @@
   "main": "src/index.ts",
   "types": "src/index.ts",
   "dependencies": {
-    "@rallly/utils": "*",
     "@rallly/database": "*",
+    "@rallly/utils": "*",
     "@tanstack/react-query": "^4.22.0",
     "@trpc/client": "^10.13.0",
     "@trpc/next": "^10.13.0",
     "@trpc/react-query": "^10.13.0",
-    "@trpc/server": "^10.13.0"
+    "@trpc/server": "^10.13.0",
+    "iron-session": "^6.3.1"
   }
 }
