--- packages/backend/src/index.ts
@@ -1,1 +1,0 @@
-export * from "./utils/token";
