+++ packages/backend/trpc/routers/polls.ts
@@ -6,7 +6,7 @@ import dayjs from "dayjs";
 import { z } from "zod";
 
 import { nanoid } from "../../utils/nanoid";
-import { possiblyPublicProcedure, publicProcedure, router } from "../trpc";
+import { possiblyPublicProcedure, publicProcedure, router } from "..";
 import { comments } from "./polls/comments";
 import { demo } from "./polls/demo";
 import { participants } from "./polls/participants";
