+++ packages/backend/trpc/routers/polls/demo.ts
@@ -2,7 +2,7 @@ import { prisma, VoteType } from "@rallly/database";
 import dayjs from "dayjs";
 
 import { nanoid } from "../../../utils/nanoid";
-import { possiblyPublicProcedure, router } from "../../trpc";
+import { possiblyPublicProcedure, router } from "../..";
 
 const participantData: Array<{ name: string; votes: VoteType[] }> = [
   {
